refactor(rust): move weak carriers under types
