+++ rust/fory-core/src/serializer/weak.rs
@@ -15,313 +15,16 @@
 // specific language governing permissions and limitations
 // under the License.
 
-//! Weak pointer serialization support for `Rc` and `Arc`.
-//!
-//! This module provides [`RcWeak<T>`] and [`ArcWeak<T>`] wrapper types that integrate
-//! Rust's `std::rc::Weak` / `std::sync::Weak` into the Fory serialization framework,
-//! with full support for:
-//
-//! - **Reference identity tracking** — weak pointers serialize as references to their
-//!   corresponding strong pointers, ensuring shared and circular references in the graph
-//!   are preserved without duplication.
-//! - **Null weak pointers** — if the strong pointer has been dropped or was never set,
-//!   the weak will serialize as a `Null` flag.
-//! - **Forward references during deserialization** — if the strong pointer appears later
-//!   in the serialized data, the weak will be resolved after deserialization using
-//!   `RefReader` callbacks.
-//!
-//! ## When to use
-//!
-//! Use these wrappers when your graph structure contains parent/child relationships
-//! or other shared edges where a strong pointer would cause a reference cycle.
-//! Storing a weak pointer avoids owning the target strongly, but serialization
-//! will preserve the link by reference ID.
-//!
-//! ## Example — Parent/Child Graph
-//!
-//! ```rust,ignore
-//! use fory_core::RcWeak;
-//! use fory_core::Fory;
-//! use std::cell::RefCell;
-//! use std::rc::Rc;
-//! use fory_derive::ForyObject;
-//!
-//! #[derive(ForyObject)]
-//! struct Node {
-//!     value: i32,
-//!     parent: RcWeak<RefCell<Node>>,
-//!     children: Vec<Rc<RefCell<Node>>>,
-//! }
-//!
-//! let mut fory = Fory::default();
-//! fory.register::<Node>(2000);
-//!
-//! let parent = Rc::new(RefCell::new(Node {
-//!     value: 1,
-//!     parent: RcWeak::new(),
-//!     children: vec![],
-//! }));
-//!
-//! let child1 = Rc::new(RefCell::new(Node {
-//!     value: 2,
-//!     parent: RcWeak::from(&parent),
-//!     children: vec![],
-//! }));
-//!
-//! let child2 = Rc::new(RefCell::new(Node {
-//!     value: 3,
-//!     parent: RcWeak::from(&parent),
-//!     children: vec![],
-//! }));
-//!
-//! parent.borrow_mut().children.push(child1);
-//! parent.borrow_mut().children.push(child2);
-//!
-//! let serialized = fory.serialize(&parent);
-//! let deserialized: Rc<RefCell<Node>> = fory.deserialize(&serialized).unwrap();
-//!
-//! assert_eq!(deserialized.borrow().children.len(), 2);
-//! for child in &deserialized.borrow().children {
-//!     let upgraded_parent = child.borrow().parent.upgrade().unwrap();
-//!     assert!(Rc::ptr_eq(&deserialized, &upgraded_parent));
-//! }
-//! ```
-//!
-//! ## Example — Arc for Multi-Threaded Graphs
-//!
-//! ```rust,ignore
-//! use fory_core::ArcWeak;
-//! use fory_core::Fory;
-//! use std::sync::{Arc, Mutex};
-//! use fory_derive::ForyObject;
-//!
-//! #[derive(ForyObject)]
-//! struct Node {
-//!     value: i32,
-//!     parent: ArcWeak<Mutex<Node>>,
-//! }
-//!
-//! let mut fory = Fory::default();
-//! fory.register::<Node>(2001);
-//!
-//! let parent = Arc::new(Mutex::new(Node { value: 1, parent: ArcWeak::new() }));
-//! let child = Arc::new(Mutex::new(Node { value: 2, parent: ArcWeak::from(&parent) }));
-//!
-//! let serialized = fory.serialize(&child);
-//! let deserialized: Arc<Mutex<Node>> = fory.deserialize(&serialized).unwrap();
-//! assert_eq!(deserialized.lock().unwrap().value, 2);
-//! ```
-//!
-//! ## Notes
-//!
-//! - These types share the same `UnsafeCell` across clones, so updating a weak in one clone
-//!   will update all of them.
-//! - During serialization, weak pointers **never serialize the target object's data directly**
-//!   — they only emit a reference to the already-serialized strong pointer, or `Null`.
-//! - During deserialization, unresolved references will be patched up by `RefReader::add_callback`
-//!   once the strong pointer becomes available.
+//! Serialization support for [`crate::types::weak::RcWeak`] and [`crate::types::weak::ArcWeak`].
 
 use crate::error::Error;
 use crate::resolver::context::{ReadContext, WriteContext};
 use crate::resolver::type_resolver::{TypeInfo, TypeResolver};
 use crate::serializer::{ForyDefault, Serializer};
+use crate::types::{ArcWeak, RcWeak};
 use crate::wire::{RefFlag, RefMode, TypeId};
-use std::cell::UnsafeCell;
 use std::rc::Rc;
 use std::sync::Arc;
-
-/// A serializable wrapper around `std::rc::Weak<T>`.
-///
-/// `RcWeak<T>` is designed for use in graph-like structures where nodes may need to hold
-/// non-owning references to other nodes (e.g., parent pointers), and you still want them
-/// to round-trip through serialization while preserving reference identity.
-///
-/// Unlike a raw `Weak<T>`, cloning `RcWeak` keeps all clones pointing to the same
-/// internal `UnsafeCell`, so updates via deserialization callbacks affect all copies.
-///
-/// # Example
-/// See module-level docs for a complete graph example.
-///
-/// # Null handling
-/// If the target `Rc<T>` has been dropped or never assigned, `upgrade()` returns `None`
-/// and serialization will write a `RefFlag::Null` instead of a reference ID.
-pub struct RcWeak<T: ?Sized> {
-    // Use Rc<UnsafeCell> so that clones share the same cell
-    inner: Rc<UnsafeCell<std::rc::Weak<T>>>,
-}
-
-impl<T: ?Sized> std::fmt::Debug for RcWeak<T> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("RcWeak")
-            .field("strong_count", &self.strong_count())
-            .field("weak_count", &self.weak_count())
-            .finish()
-    }
-}
-
-impl<T> RcWeak<T> {
-    pub fn new() -> Self {
-        RcWeak {
-            inner: Rc::new(UnsafeCell::new(std::rc::Weak::new())),
-        }
-    }
-}
-
-impl<T: ?Sized> RcWeak<T> {
-    pub fn upgrade(&self) -> Option<Rc<T>> {
-        unsafe { (*self.inner.get()).upgrade() }
-    }
-
-    pub fn strong_count(&self) -> usize {
-        unsafe { (*self.inner.get()).strong_count() }
-    }
-
-    pub fn weak_count(&self) -> usize {
-        unsafe { (*self.inner.get()).weak_count() }
-    }
-
-    pub fn ptr_eq(&self, other: &Self) -> bool {
-        unsafe { std::rc::Weak::ptr_eq(&*self.inner.get(), &*other.inner.get()) }
-    }
-
-    pub fn update(&self, weak: std::rc::Weak<T>) {
-        unsafe {
-            *self.inner.get() = weak;
-        }
-    }
-
-    pub fn from_std(weak: std::rc::Weak<T>) -> Self {
-        RcWeak {
-            inner: Rc::new(UnsafeCell::new(weak)),
-        }
-    }
-}
-
-impl<T: ?Sized> PartialEq for RcWeak<T> {
-    fn eq(&self, other: &Self) -> bool {
-        self.ptr_eq(other)
-    }
-}
-
-impl<T: ?Sized> Eq for RcWeak<T> {}
-
-impl<T> Default for RcWeak<T> {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl<T: ?Sized> Clone for RcWeak<T> {
-    fn clone(&self) -> Self {
-        // Clone the Rc, not the inner Weak - this way clones share the same cell!
-        RcWeak {
-            inner: self.inner.clone(),
-        }
-    }
-}
-
-impl<T: ?Sized> From<&Rc<T>> for RcWeak<T> {
-    fn from(rc: &Rc<T>) -> Self {
-        RcWeak::from_std(Rc::downgrade(rc))
-    }
-}
-
-unsafe impl<T: ?Sized> Send for RcWeak<T> where std::rc::Weak<T>: Send {}
-unsafe impl<T: ?Sized> Sync for RcWeak<T> where std::rc::Weak<T>: Sync {}
-
-/// A serializable wrapper around `std::sync::Weak<T>` (thread-safe).
-///
-/// `ArcWeak<T>` works exactly like [`RcWeak<T>`] but is intended for use with
-/// multi-threaded shared graphs where strong pointers are `Arc<T>`.
-///
-/// All clones of an `ArcWeak<T>` share the same `UnsafeCell` so deserialization
-/// updates propagate to all copies.
-///
-/// # Example
-/// See module-level docs for an `Arc<Mutex<Node>>` usage example.
-pub struct ArcWeak<T: ?Sized> {
-    // Use Arc<UnsafeCell> so that clones share the same cell
-    inner: Arc<UnsafeCell<std::sync::Weak<T>>>,
-}
-
-impl<T: ?Sized> std::fmt::Debug for ArcWeak<T> {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        f.debug_struct("ArcWeak")
-            .field("strong_count", &self.strong_count())
-            .field("weak_count", &self.weak_count())
-            .finish()
-    }
-}
-
-impl<T> ArcWeak<T> {
-    pub fn new() -> Self {
-        ArcWeak {
-            inner: Arc::new(UnsafeCell::new(std::sync::Weak::new())),
-        }
-    }
-}
-
-impl<T: ?Sized> ArcWeak<T> {
-    pub fn upgrade(&self) -> Option<Arc<T>> {
-        unsafe { (*self.inner.get()).upgrade() }
-    }
-
-    pub fn strong_count(&self) -> usize {
-        unsafe { (*self.inner.get()).strong_count() }
-    }
-
-    pub fn weak_count(&self) -> usize {
-        unsafe { (*self.inner.get()).weak_count() }
-    }
-
-    pub fn ptr_eq(&self, other: &Self) -> bool {
-        unsafe { std::sync::Weak::ptr_eq(&*self.inner.get(), &*other.inner.get()) }
-    }
-
-    pub fn update(&self, weak: std::sync::Weak<T>) {
-        unsafe {
-            *self.inner.get() = weak;
-        }
-    }
-
-    pub fn from_std(weak: std::sync::Weak<T>) -> Self {
-        ArcWeak {
-            inner: Arc::new(UnsafeCell::new(weak)),
-        }
-    }
-}
-
-impl<T: ?Sized> PartialEq for ArcWeak<T> {
-    fn eq(&self, other: &Self) -> bool {
-        self.ptr_eq(other)
-    }
-}
-
-impl<T: ?Sized> Eq for ArcWeak<T> {}
-
-impl<T> Default for ArcWeak<T> {
-    fn default() -> Self {
-        Self::new()
-    }
-}
-
-impl<T: ?Sized> Clone for ArcWeak<T> {
-    fn clone(&self) -> Self {
-        // Clone the Arc, not the inner Weak - this way clones share the same cell!
-        ArcWeak {
-            inner: self.inner.clone(),
-        }
-    }
-}
-
-impl<T: ?Sized> From<&Arc<T>> for ArcWeak<T> {
-    fn from(arc: &Arc<T>) -> Self {
-        ArcWeak::from_std(Arc::downgrade(arc))
-    }
-}
-
-unsafe impl<T: ?Sized + Send + Sync> Send for ArcWeak<T> {}
-unsafe impl<T: ?Sized + Send + Sync> Sync for ArcWeak<T> {}
 
 impl<T: Serializer + ForyDefault + 'static> Serializer for RcWeak<T> {
     fn fory_is_shared_ref() -> bool {
@@ -651,13 +354,10 @@ fn read_arc_weak<T: Serializer + ForyDefault + 'static>(
             if let Some(arc) = context.ref_reader.get_arc_ref::<T>(ref_id) {
                 weak.update(Arc::downgrade(&arc));
             } else {
-                // Capture the raw pointer to the UnsafeCell so we can update it in the callback
-                let weak_ptr = weak.inner.get();
+                let callback_weak = weak.clone();
                 context.ref_reader.add_callback(Box::new(move |ref_reader| {
                     if let Some(arc) = ref_reader.get_arc_ref::<T>(ref_id) {
-                        unsafe {
-                            *weak_ptr = Arc::downgrade(&arc);
-                        }
+                        callback_weak.update(Arc::downgrade(&arc));
                     }
                 }));
             }
